fix: resolve home page cache issues and harmonize animated icons across navbar and footer
- Added revalidatePath('/') to project CRUD actions to ensure instant public updates.
- Extracted shared AnimatedIcons component for Lucide icons (Home, User, Projects, Send).
- Fixed mobile navigation issues by using smoothScrollTo and delaying menu closure.
- Synchronized Footer navigation icons with Navbar's animated Lucide style.
